feat(delete): Move account delete route handler to account delete manager class
